+++ src/main/java/gim/logic/parser/AddCommandParser.java
@@ -16,7 +16,7 @@ import gim.model.exercise.Name;
 import gim.model.exercise.Reps;
 import gim.model.exercise.Sets;
 import gim.model.exercise.Weight;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 
 
 /**
