+++ src/main/java/gim/logic/parser/ParserUtil.java
@@ -9,7 +9,7 @@ import gim.model.exercise.Name;
 import gim.model.exercise.Reps;
 import gim.model.exercise.Sets;
 import gim.model.exercise.Weight;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
