+++ src/main/java/gim/model/date/Date.java
@@ -1,4 +1,4 @@
-package gim.model.tag;
+package gim.model.date;
 
 import static gim.commons.util.AppUtil.checkArgument;
 import static java.util.Objects.requireNonNull;
