+++ src/main/java/gim/model/util/SampleDataUtil.java
@@ -7,7 +7,7 @@ import gim.model.exercise.Name;
 import gim.model.exercise.Reps;
 import gim.model.exercise.Sets;
 import gim.model.exercise.Weight;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 
 /**
  * Contains utility methods for populating {@code ExerciseTracker} with sample data.
