+++ src/test/java/gim/logic/parser/EditCommandParserTest.java
@@ -38,7 +38,7 @@ import gim.model.exercise.Name;
 import gim.model.exercise.Reps;
 import gim.model.exercise.Sets;
 import gim.model.exercise.Weight;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 import gim.testutil.EditExerciseDescriptorBuilder;
