+++ src/test/java/gim/logic/parser/ParserUtilTest.java
@@ -12,7 +12,7 @@ import gim.model.exercise.Name;
 import gim.model.exercise.Reps;
 import gim.model.exercise.Sets;
 import gim.model.exercise.Weight;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 
 
 public class ParserUtilTest {
