+++ src/test/java/gim/testutil/ExerciseBuilder.java
@@ -8,7 +8,7 @@ import gim.model.exercise.Name;
 import gim.model.exercise.Reps;
 import gim.model.exercise.Sets;
 import gim.model.exercise.Weight;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 import gim.model.util.SampleDataUtil;
