+++ src/test/java/gim/testutil/ExerciseUtil.java
@@ -9,7 +9,7 @@ import static gim.logic.parser.CliSyntax.PREFIX_WEIGHT;
 import gim.logic.commands.AddCommand;
 import gim.logic.commands.EditCommand.EditExerciseDescriptor;
 import gim.model.exercise.Exercise;
-import gim.model.tag.Date;
+import gim.model.date.Date;
 
 
 /**
